[#IC-448] (+) move message activity type to types.ts
