+++ HandleNHNotificationCall/handler.ts
@@ -18,10 +18,7 @@ import { KindEnum as NotifyKind } from "../generated/notifications/NotifyMessage
 
 import { OrchestratorName as CreateOrUpdateInstallationOrchestrator } from "../HandleNHCreateOrUpdateInstallationCallOrchestrator/handler";
 import { OrchestratorName as DeleteInstallationOrchestratorName } from "../HandleNHDeleteInstallationCallOrchestrator/handler";
-import {
-  NhNotifyMessageRequest,
-  NhTarget
-} from "../HandleNHNotifyMessageCallActivityQueue/handle";
+import { NhNotifyMessageRequest, NhTarget } from "../utils/types";
 
 export const NotificationMessage = t.union([
   NotifyMessage,
